backend/requirements.txt
Django==5.2.3
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
python-decouple==3.8
psycopg2-binary==2.9.10
yookassa==3.9.0
openpyxl==3.1.5
